BUG: remove deprecated `np.float` dtype

+++ glue/core/data_factories/tests/test_data_factories.py
@@ -106,7 +106,7 @@ def test_csv_pandas_factory():
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname, factory=df.pandas_read_table)
     assert d['a'].dtype == np.int64
-    assert d['b'].dtype == np.float
+    assert d['b'].dtype == float
     assert d['c'].dtype.kind == 'U'
     cat_comp = d.find_component_id('c')
     assert isinstance(d.get_component(cat_comp), CategoricalComponent)
@@ -129,7 +129,7 @@ def test_dtype_float():
     data = b'# a, b\n1., 1 \n2, 2 \n3, 3'
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname)
-    assert d['a'].dtype == np.float
+    assert d['a'].dtype == float
 
 
 def test_dtype_str_on_categorical():
@@ -143,7 +143,7 @@ def test_dtype_badtext():
     data = b'# a, b\nlabel1, 1 \n2, 2 \n3, 3\n4, 4\n5, 5\n6, 6'
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname)
-    assert d['a'].dtype == np.float
+    assert d['a'].dtype == float
     assert_array_equal(d['a'], [np.nan, 2, 3, 4, 5, 6])
 
 
@@ -151,7 +151,7 @@ def test_dtype_missing_data_col2():
     data = b'# a, b\n1 , 1 \n2,  \n3, 3.0'
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname)
-    assert d['b'].dtype == np.float
+    assert d['b'].dtype == float
     assert_array_equal(d['b'], [1, np.nan, 3])
 
 
@@ -159,7 +159,7 @@ def test_dtype_missing_data_col1():
     data = b'# a, b\n1.0, 1 \n , 2 \n3, 3'
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname)
-    assert d['a'].dtype == np.float
+    assert d['a'].dtype == float
     assert_array_equal(d['a'], [1, np.nan, 3])
 
 
@@ -167,7 +167,7 @@ def test_column_spaces():
     data = b'#a, b\nhere I go, 1\n2, 3\n3, 4\n5, 6\n7, 8'
     with make_file(data, '.csv') as fname:
         d = df.load_data(fname)
-    assert d['a'].dtype == np.float
+    assert d['a'].dtype == float
     assert_array_equal(d['a'], [np.nan, 2, 3, 5, 7])
